Fixed bug where status message did not update
Previously, the statusDiv was not updating
correctly after each round because the getResult
function did not return the correct value.
Specifically, the getResult function failed to
return the appropriate values :
(1 for win, -1 for loss, 0 for tie).

Changes made:
- Now getResult function always return a value
Since result variable is not undefined anymore,
the correct status message is displayed.